Use size_t as parameter type in the getenv member of struct sapi_module
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@30603 c90b9560-bf6c-de11-be94-00142212c4b1

+++ main/SAPI.c
@@ -570,7 +570,7 @@ SAPI_API struct stat *sapi_get_stat()
 }
 
 
-SAPI_API char *sapi_getenv(char *name, int name_len)
+SAPI_API char *sapi_getenv(char *name, size_t name_len)
 {
 	if (sapi_module.getenv) {
 		SLS_FETCH();
